requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Werkzeug==0.10.4
aniso8601==1.0.0
itsdangerous==0.24
pycrypto==2.6
pytz==2015.6
requests==2.8.1
six==1.10.0
